added generating reports - fixing stuff v2
